type change for vertex texcoords, now an int

+++ src/display/model/mesh/mesh.h
@@ -13,7 +13,7 @@
 typedef struct {
     glm::vec3 Position;
     glm::vec3 Normal;
-    glm::vec2 TexCoords;
+    glm::ivec2 TexCoords;
 } VERTEX;
 
 typedef struct {
